REAL in SQLite is always a double
stored as an 8-byte IEEE floating point number
https://www.sqlite.org/datatype3.html

+++ src/MeilisearchItem.cs
@@ -15,7 +15,7 @@ public record MeilisearchItem(
     string[]? Studios,
     string[]? Tags,
     bool? IsFolder,
-    float? CommunityRating,
-    float? CriticRating,
+    double? CommunityRating,
+    double? CriticRating,
     string? Path
 );
